Add EnhancedRiskManagement component and refactor BigGradientBox styles
