+++ apps/earn-protocol-landing-page/components/layout/LandingPageContent/components/BigGradientBox.tsx
@@ -1,9 +1,19 @@
 import { type FC, type ReactNode } from 'react'
+import clsx from 'clsx'
 
 import bigGradientBoxStyles from './BigGradientBox.module.scss'
 
 export const BigGradientBox: FC<{
   children: ReactNode
-}> = ({ children }) => {
-  return <div className={bigGradientBoxStyles.bigGradientBox}>{children}</div>
+  reversed?: boolean
+}> = ({ children, reversed = false }) => {
+  return (
+    <div
+      className={clsx(bigGradientBoxStyles.bigGradientBox, {
+        [bigGradientBoxStyles.reversed]: reversed,
+      })}
+    >
+      {children}
+    </div>
+  )
 }
