+++ apps/earn-protocol-landing-page/components/layout/LandingPageContent/components/LandingPageHero.module.scss
