+++ apps/earn-protocol-landing-page/components/layout/LandingPageContent/content/EnhancedRiskManagement.module.scss
@@ -1,0 +1,8 @@
+.enhancedRiskManagementHeaderWrapper {
+  text-align: center;
+  padding: 0 var(--spacing-space-3x-large);
+}
+
+.enhancedRiskManagementHeader {
+  font-weight: 600;
+}
