+++ apps/earn-protocol-landing-page/components/layout/LandingPageContent/index.ts
@@ -1,10 +1,11 @@
-export { LandingPageHero } from '@/components/layout/LandingPageContent/LandingPageHero'
 export { BigGradientBox } from '@/components/layout/LandingPageContent/components/BigGradientBox'
 export { Emphasis } from '@/components/layout/LandingPageContent/components/Emphasis'
-export { ProtocolScroller } from '@/components/layout/LandingPageContent/components/ProtocolScroller'
+export { LandingPageHero } from '@/components/layout/LandingPageContent/components/LandingPageHero'
 export { MarketingPoints } from '@/components/layout/LandingPageContent/components/MarketingPoints'
+export { ProtocolScroller } from '@/components/layout/LandingPageContent/components/ProtocolScroller'
 
+export { EffortlessAccessBlock } from '@/components/layout/LandingPageContent/content/EffortlessAccessBlock'
+export { EnhancedRiskManagement } from '@/components/layout/LandingPageContent/content/EnhancedRiskManagement'
 export { HigherYieldsBlock } from '@/components/layout/LandingPageContent/content/HigherYieldsBlock'
 export { SummerFiProBox } from '@/components/layout/LandingPageContent/content/SummerFiProBox'
-export { EffortlessAccessBlock } from '@/components/layout/LandingPageContent/content/EffortlessAccessBlock'
 export { SupportedNetworksList } from '@/components/layout/LandingPageContent/content/SupportedNetworksList'
